Removing the file_get_contents

+++ classes/class-lsx-wetu-importer-banner-integration.php
@@ -231,51 +231,54 @@ class LSX_WETU_Importer_Banner_Integration extends LSX_WETU_Importer {
 
 	public function attach_external_image2( $url = null, $post_data = array(), $post_id = '' ) {
 		if ( ! $url ) { return new WP_Error( 'missing', 'Need a valid URL' ); }
+		$att_id = false;
 
 		require_once( ABSPATH . 'wp-admin/includes/file.php' );
 		require_once( ABSPATH . 'wp-admin/includes/media.php' );
 		require_once( ABSPATH . 'wp-admin/includes/image.php' );
 
-		$tmp = tempnam( '/tmp', 'FOO' );
-		$image = file_get_contents( $url );
-		file_put_contents( $tmp, $image );
-		chmod( $tmp, '777' );
+		$tmp   = tempnam( '/tmp', 'FOO' );
+		$image = wp_remote_get( $url );
 
-		preg_match( '/[^\?]+\.(tif|TIFF|jpg|JPG|jpe|JPE|jpeg|JPEG|gif|GIF|png|PNG|pdf|PDF|bmp|BMP)/', $url, $matches );
-		$url_filename = basename( $matches[0] );
-		$url_filename = str_replace( '%20','_', $url_filename );
-		// extract filename from url for title.
-		$url_type = wp_check_filetype( $url_filename );
-
-		// assemble file data (should be built like $_FILES since wp_handle_sideload() will be using).
-		$file_array['tmp_name'] = $tmp;
-
-		if ( ! empty( $filename ) && ' ' != $filename ) {
-			$file_array['name'] = $filename . '.' . $url_type['ext'];
-		} else {
-			$file_array['name'] = $url_filename;
+		if ( ! empty( $image ) && isset( $image['response'] ) && isset( $image['response']['code'] ) && 200 === $image['response']['code'] ) {
+			file_put_contents( $tmp, $image['body'] );
+			chmod( $tmp, '777' );
+	
+			preg_match( '/[^\?]+\.(tif|TIFF|jpg|JPG|jpe|JPE|jpeg|JPEG|gif|GIF|png|PNG|pdf|PDF|bmp|BMP)/', $url, $matches );
+			$url_filename = basename( $matches[0] );
+			$url_filename = str_replace( '%20','_', $url_filename );
+			// extract filename from url for title.
+			$url_type = wp_check_filetype( $url_filename );
+	
+			// assemble file data (should be built like $_FILES since wp_handle_sideload() will be using).
+			$file_array['tmp_name'] = $tmp;
+	
+			if ( ! empty( $filename ) && ' ' != $filename ) {
+				$file_array['name'] = $filename . '.' . $url_type['ext'];
+			} else {
+				$file_array['name'] = $url_filename;
+			}
+	
+			// set additional wp_posts columns.
+			if ( empty( $post_data['post_title'] ) ) {
+				$url_filename = str_replace( '%20', ' ', $url_filename );
+				$post_data['post_title'] = basename( $url_filename, '.' . $url_type['ext'] );
+			}
+	
+			// make sure gets tied to parent.
+			if ( empty( $post_data['post_parent'] ) ) {
+				$post_data['post_parent'] = $post_id;
+			}
+	
+			// do the validation and storage stuff.
+			$att_id = media_handle_sideload( $file_array, $post_id, null, $post_data );
+	
+			// If error storing permanently, unlink.
+			if ( is_wp_error( $att_id ) ) {
+				unlink( $file_array['tmp_name'] );
+				return false;
+			}
 		}
-
-		// set additional wp_posts columns.
-		if ( empty( $post_data['post_title'] ) ) {
-			$url_filename = str_replace( '%20', ' ', $url_filename );
-			$post_data['post_title'] = basename( $url_filename, '.' . $url_type['ext'] );
-		}
-
-		// make sure gets tied to parent.
-		if ( empty( $post_data['post_parent'] ) ) {
-			$post_data['post_parent'] = $post_id;
-		}
-
-		// do the validation and storage stuff.
-		$att_id = media_handle_sideload( $file_array, $post_id, null, $post_data );
-
-		// If error storing permanently, unlink.
-		if ( is_wp_error( $att_id ) ) {
-			unlink( $file_array['tmp_name'] );
-			return false;
-		}
-
 		return $att_id;
 	}
 }
